implemented Movable abstract class
