+++ Assets/Scripts/EnemyController.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 using UnityEngine.AI;
 using Utils;
 
-public class EnemyController : MonoBehaviour, IMovable
+public class EnemyController : Movable
 {
     [Tooltip("The area the enemy wanders in.")]
     [SerializeField]
@@ -23,14 +23,6 @@ public class EnemyController : MonoBehaviour, IMovable
     [Tooltip("The ground the NPC should walk on.")]
     [SerializeField]
     private LayerMask moveLayer;
-
-    [Tooltip("The speed the NPC rotates with.")]
-    [SerializeField]
-    private float rotationSpeed;
-
-    [Tooltip("The maximum difference in degrees for the NPC between look direction and target direction in order to be facing the target.")]
-    [SerializeField]
-    private float rotationTolerance;
 
     [Tooltip("The speed the NPC chases targets with.")]
     [SerializeField]
@@ -53,7 +45,6 @@ public class EnemyController : MonoBehaviour, IMovable
     private bool showEscapeRadius;
 
     // component references
-    private NavMeshAgent _agent;
     private WanderAI _wanderAI;
     private AttackLogic _attackLogic;
 
@@ -64,10 +55,10 @@ public class EnemyController : MonoBehaviour, IMovable
     private float _initialSpeed;
 
 
-    private void Awake()
+    protected override void Awake()
     {
+        base.Awake();
         // init components
-        _agent = GetComponent<NavMeshAgent>();
         _wanderAI = new WanderAI();
         _attackLogic = GetComponent<AttackLogic>();
         if (_attackLogic == null)
@@ -88,7 +79,7 @@ public class EnemyController : MonoBehaviour, IMovable
             wanderArea.FreezeArea();
         _timer = wanderTimer;
         // remember initial speed
-        _initialSpeed = _agent.speed;
+        _initialSpeed = _navMeshAgent.speed;
     }
 
     /// <summary>
@@ -141,7 +132,7 @@ public class EnemyController : MonoBehaviour, IMovable
     private void StartChasing(GameObject target)
     {
         _isChasing = true;
-        _agent.speed = runningSpeed;
+        _navMeshAgent.speed = runningSpeed;
         _attackLogic.StartAttack(target);
     }
 
@@ -151,7 +142,7 @@ public class EnemyController : MonoBehaviour, IMovable
     private void StopChasing()
     {
         _isChasing = false;
-        _agent.speed = _initialSpeed;
+        _navMeshAgent.speed = _initialSpeed;
         _attackLogic.StopAttack();
     }
 
@@ -170,48 +161,10 @@ public class EnemyController : MonoBehaviour, IMovable
         else
         {
             // don't use SetDestination but set calculated path
-            NavMeshPath wanderPath = _wanderAI.GetNextWanderPath(wanderArea, _agent, moveLayer.value);
-            _agent.SetPath(wanderPath);
-            _agent.isStopped = false;
+            NavMeshPath wanderPath = _wanderAI.GetNextWanderPath(wanderArea, _navMeshAgent, moveLayer.value);
+            _navMeshAgent.SetPath(wanderPath);
+            _navMeshAgent.isStopped = false;
         }
-    }
-
-    // ----- Note: same in PlayerController --> make IMovable abstract class and inherit? ------
-
-    /// <summary>
-    /// Faces the target. Returns true if facing the target.
-    /// </summary>
-    /// <param name="target">The target to face</param>
-    /// <param name="shouldTurn">Whether the object should turn to the target or not</param>
-    /// <param name="difference">The difference between object and target in degrees</param>
-    /// <returns>Whether the object is facing the target</returns>
-    public bool FaceTarget(GameObject target, bool shouldTurn, out float difference)
-    {
-        Vector3 direction = (target.transform.position - transform.position).normalized;
-        Quaternion lookRotation = Quaternion.LookRotation(new Vector3(direction.x, 0f, direction.z));
-        difference = Mathf.Abs(lookRotation.eulerAngles.magnitude - transform.rotation.eulerAngles.magnitude);
-        bool facesTarget = difference < rotationTolerance;
-        if (!facesTarget && shouldTurn)
-            transform.rotation = Quaternion.Slerp(transform.rotation, lookRotation, Time.deltaTime * rotationSpeed);
-        return facesTarget;
-    }
-
-    /// <summary>
-    /// Sets destination for NavMesh agent and starts moving.
-    /// </summary>
-    /// <param name="position">The position to move to</param>
-    public void Move(Vector3 position)
-    {
-        _agent.SetDestination(position);
-        _agent.isStopped = false;
-    }
-
-    /// <summary>
-    /// Stops moving.
-    /// </summary>
-    public void StopMoving()
-    {
-        _agent.isStopped = true;
     }
 
     /// <summary>
