+++ Assets/Scripts/Player/PlayerController.cs
@@ -10,9 +10,9 @@ using Utils;
 
 namespace Player
 {
-    [RequireComponent(typeof(NavMeshAgent))]
+    
     [RequireComponent(typeof(PlayerInput))]
-    public class PlayerController : MonoBehaviour, IMovable
+    public class PlayerController : Movable
     {
         [Tooltip("The clickable layer. Defines where the player can click/move")] [SerializeField]
         private LayerMask moveClickLayers;
@@ -36,14 +36,6 @@ namespace Player
         [Tooltip("The min- and max-distance of the camera")]
         [SerializeField]
         private Range cameraDistanceRange;
-
-        [Tooltip("The speed the player rotates with")]
-        [SerializeField]
-        private int rotationSpeed;
-
-        [Tooltip("The maximum difference in degrees for the player between look direction and target direction in order to be facing the target.")]
-        [SerializeField]
-        private int rotationTolerance;
         
         [Tooltip("The inventory UI to toggle when pressing the inventory key")]
         [SerializeField]
@@ -53,7 +45,6 @@ namespace Player
         public static event PlayerPositionChanged OnPlayerPositionUpdated;
         
         //Component references
-        private NavMeshAgent _navMeshAgent;
         private Camera _camera;
         private Controls _controls;
         private AttackLogic _attackLogic;
@@ -71,9 +62,9 @@ namespace Player
         /// <summary>
         /// Gets references and sets up the controls.
         /// </summary>
-        private void Awake()
+        protected override void Awake()
         {
-            _navMeshAgent = GetComponent<NavMeshAgent>();
+            base.Awake();
             _attackLogic = GetComponent<AttackLogic>();
             SetUpControls();
             PauseMenu.OnPause += OnPause;
@@ -164,58 +155,16 @@ namespace Player
         }
 
         /// <summary>
-        /// Player moves to the pressed location using the NavMeshAgent.
+        /// Player moves to the pressed location using the NavMeshAgent. Click effect is created.
         /// <param name="hit">The point on the ground the player should move to.</param>
         /// </summary>
         private void Move(RaycastHit hit)
         {
-            // Set destination for nav mesh agent
-            _navMeshAgent.SetDestination(hit.point);
-            _navMeshAgent.isStopped = false;
-
+            base.Move(hit.point);
             // Create click point effect
             Instantiate(clickEffect, hit.point + Vector3.up * 5, Quaternion.identity);
 
 			OnPlayerPositionUpdated?.Invoke(transform.position);
-        }
-
-        // ----- Note: same in EnemyController --> make IMovable abstract class and inherit? ------
-
-        /// <summary>
-        /// Faces the target. Returns true if facing the target.
-        /// </summary>
-        /// <param name="target">The target to face</param>
-        /// <param name="shouldTurn">Whether the object should turn to the target or not</param>
-        /// <param name="difference">The difference between object and target in degrees</param>
-        /// <returns>Whether the object is facing the target</returns>
-        public bool FaceTarget(GameObject target, bool shouldTurn, out float difference)
-        {
-            Vector3 direction = (target.transform.position - transform.position).normalized;
-            Quaternion lookRotation = Quaternion.LookRotation(new Vector3(direction.x, 0f, direction.z));
-            difference = Mathf.Abs(lookRotation.eulerAngles.magnitude - transform.rotation.eulerAngles.magnitude);
-            bool facesTarget = difference < rotationTolerance;
-            if (!facesTarget && shouldTurn)
-                transform.rotation = Quaternion.Slerp(transform.rotation, lookRotation, Time.deltaTime * rotationSpeed);
-            return facesTarget;
-        }
-
-        /// <summary>
-        /// Move to a certain position.
-        /// </summary>
-        /// <param name="position">The position to move to</param>
-        public void Move(Vector3 position)
-        {
-            // Set destination for nav mesh agent
-            _navMeshAgent.SetDestination(position);
-            _navMeshAgent.isStopped = false;
-        }
-
-        /// <summary>
-        /// Stops moving.
-        /// </summary>
-        public void StopMoving()
-        {
-            _navMeshAgent.isStopped = true;
         }
 
         /// <summary>
